The integration test for the New Launch Instance wizard
The possibility to create a new instance by means of the Angular
Launch Instance wizard is tested, then it is verified that this
instance is deletable as well. During the launch the same default
flavor and image as in case of legacy wizard are used. It is expected
that 'private' / 'public' network will be available to attach to the
instance being created by regular / admin user, correspondingly.

Co-Authored-By: Sergei Chipiga <schipiga@mirantis.com>
Co-Authored-By: Timur Sufiev <tsufiev@mirantis.com>

Implements blueprint: horizon-integration-tests-coverage
Change-Id: I99ee271c5294c842147e29766f6f459becf8686a

Conflicts:
	openstack_dashboard/test/integration_tests/regions/menus.py
	openstack_dashboard/test/integration_tests/tests/test_instances.py