After submit modal must be absent if check is true
Need to wait until modal background is absent after submit form if
check is true, commonly. It is a part of check: in valid case after
submit, modal is fade (except some cases where it will be omitted).

If in check notification is closed, it doesn't need to wait modal is
fade, because `close_notification` method waits it inside.